third_party/blink/web_tests/platform/generic/fast/forms/focus-selection-input-expected.txt

EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
<input type="text"> focus selection

This test checks whether the selection is restored, cleared, or set to the full range when using different ways to focus a text field. These results all match Mozilla, except test 6, which selects the whole field contents to match all other cases of keyboard focus. When running manually, please follow the steps below. In the test harness, the test runs automatically.

1) Hit tab, the whole phrase "SUCCESS and SUCCESS" should be selected: 
2) Hit Ctrl-Option-J (or Alt-J on Windows), the whole phrase "SUCCESS and SUCCESS" should be selected: 
3) , only the word "SUCCESS" should be selected: 
4) Click in the input field, nothing should be selected: 
5) Click this label, only the word "SUCCESS" should be selected: 
6) Hit Ctrl-Option-U (or Alt-U on Windows), the whole phrase "SUCCESS and SUCCESS" should be selected: 
7) , only the word "SUCCESS" should be selected: 
8) 
, only the word "SUCCESS" should be selected:

9)
Hit Ctrl-Option-I (or Alt-I on Windows), only the word "SUCCESS" should be selected:



PASS successfullyParsed is true

TEST COMPLETE

PASS first.selectionStart is 12
PASS first.selectionEnd is 19
PASS first.selectionStart is 0
PASS first.selectionEnd is 19
PASS second.selectionStart is 11
PASS second.selectionEnd is 18
PASS second.selectionStart is 11
PASS second.selectionEnd is 18
PASS third.selectionStart is 11
PASS third.selectionEnd is 18
PASS third.selectionStart is 11
PASS third.selectionEnd is 18
PASS fourth.selectionStart is 12
PASS fourth.selectionEnd is 19
PASS fourth.selectionStart is 0
PASS fourth.selectionEnd is 0
PASS fifth.selectionStart is 11
PASS fifth.selectionEnd is 18
PASS fifth.selectionStart is 11
PASS fifth.selectionEnd is 18
PASS sixth.selectionStart is 12
PASS sixth.selectionEnd is 19
PASS sixth.selectionStart is 0
PASS sixth.selectionEnd is 19
PASS seventh.selectionStart is 11
PASS seventh.selectionEnd is 18
PASS seventh.selectionStart is 11
PASS seventh.selectionEnd is 18
PASS eighth.selectionStart is 11
PASS eighth.selectionEnd is 18
PASS eighth.selectionStart is 11
PASS eighth.selectionEnd is 18
PASS ninth.selectionStart is 11
PASS ninth.selectionEnd is 18
PASS ninth.selectionStart is 11
PASS ninth.selectionEnd is 18
